+++ changes.md
@@ -1,1 +1,1 @@
-- Changed cslol-dll.dll download source to separate host
+- Changed mod-tools to download as a single zip file